* Se corrige test de multiple choice clasico. * Se reemplazan ciclos por métodos de stream. * Se formatea el código.

+++ src/main/java/edu/fiuba/algo3/modelo/Jugador.java
@@ -13,8 +13,8 @@ public class Jugador {
         return this.puntaje;
     }
 
-    public void actualizarPuntaje(int puntos) {
-        this.puntaje+=puntos;
+    public void actualizarPuntaje(long puntos) {
+        this.puntaje += puntos;
     }
 
     public String obtenerNombre() {
